requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
requests==2.26
python-dateutil==2.8.2
environs==9.3.5